ci(release): setup semantic-release
The `if` ensures that the release job only runs on the default branch.
https://stackoverflow.com/a/58142412

I wanted to use this plugin for the release of the plugin as well but
semantic-release doesn't support local files as plugins. That's why I
created a fake local-package that just uses this package via a relative
path. Libraries shouldn't use the `file:` installation method but it
should be fine since it is only for a dev dependency.
https://docs.npmjs.com/cli/v8/configuring-npm/package-json#local-paths

I had to update the `branches` part of the release config since it
expects `master` instead of `main` by default. I also had to overwrite
the plugins array to use this plugin to generate the release notes
instead of the default plugin.